fitness functions done

+++ governor_model/governor/fitness.h
@@ -6,6 +6,6 @@
 #define EMBEDDEDSYSTEMS_FITNESS_H
 #include "GA.h"
 
-float fitness(chromosome* c);
+double fitness(chromosome* c);
 
 #endif //EMBEDDEDSYSTEMS_FITNESS_H
